Fix: wrong URL to fontello.css throws error
I believe Cornelis was versioning the fontello folder to work around
an issue which is now resolved, so "fontello-010" in the path should
be stable. As an alternative, I tried to include all.css, which
includes the correct fontello.css, but the font in the rendered PDF
was obviously different.

Refs: https://github.com/syslabcom/scrum/issues/519